changed my PR with Ryan's thought I pushed Popup components, Popup-related func and properties to ProgressBar
