+++ lib/Slider/Slider.js
@@ -290,6 +290,21 @@ module.exports = kind(
 	*/
 	create: function () {
 		ProgressBar.prototype.create.apply(this, arguments);
+		this._nf = new ilib.NumFmt({type: 'percentage', useNative: false});
+		// FIXME: Backwards-compatibility for deprecated property - can be removed when
+		// the popupContentUpperCase property is fully deprecated and removed. The legacy
+		// property takes precedence if it exists.
+		if (this.popupContentUpperCase !== null) this.uppercase = this.popupContentUpperCase;
+		this.createComponents(this.moreComponents, {owner: this});
+		this.$.knob.createComponents(this.popupComponents, {owner: this});
+		this.initValue(); 
+		this.disabledChanged();
+		this.knobClassesChanged();
+		this.popupLabelClassesChanged();
+		this.popupContentChanged();
+		this.tapAreaClassesChanged();
+		this.initSliderStyles();
+		this.jumpIncrementChanged();
 	},
 
 	/**
